refactor: flatten the opportunity card component
