--- frontend/src/components/application/OpportunityItem.tsx
@@ -1,8 +1,0 @@
-export const OpportunityItem = ({opKey, opValue}: {opKey: string, opValue?: string | null}) => {
-    return (
-        <div>
-            <dt className="margin-right-1 text-bold">{ opKey }:</dt>
-            <dd>{ opValue ?? '--'}</dd>
-        </div>
-    )
-}
